autotests bugs corrections
