+++ scripts/Make-Client.sh
@@ -1,45 +1,51 @@
 #!/bin/bash
 
-# Input parameters
 Port=$1
 ClientPath=$2
 UIPath=$3
 OutputPath=$4
+Debug=$5
 
-# Navigate to the UIPath
-cd "$UIPath" || { echo "UIPath not found: $UIPath"; exit 1; }
-
-# Run npm build
-npm run build:cli"$Port"
-if [ $? -ne 0 ]; then
-    echo "npm build failed"
+if [ -z "$Port" ]  [ -z "$ClientPath" ]  [ -z "$UIPath" ]  [ -z "$OutputPath" ]; then
+    echo "Usage: $0 <Port> <ClientPath> <UIPath> <OutputPath> [Debug]"
     exit 1
 fi
 
-# Copy the built files
-mkdir -p "$ClientPath/static/js" "$ClientPath/static/html"
-cp "$UIPath/dist/main.js" "$ClientPath/static/js/" || { echo "Failed to copy main.js"; exit 1; }
-cp "$UIPath/dist/index.html" "$ClientPath/static/html/" || { echo "Failed to copy index.html"; exit 1; }
+cd "$UIPath"  exit 1
 
-# Navigate to the ClientPath
-cd "$ClientPath" || { echo "ClientPath not found: $ClientPath"; exit 1; }
+npm run build:cli$Port
+if [ $? -ne 0 ]; then
+    echo "Build failed"
+    exit 1
+fi
 
-# Build the Go executable
+mkdir -p "$ClientPath/static/react/js"
+cp "$UIPath/dist/main.js" "$ClientPath/static/react/js/"
+
+if [ "$Debug" == "true" ]; then
+    cp "$UIPath/dist/main.js.map" "$ClientPath/static/react/js/"
+fi
+
+mkdir -p "$ClientPath/static/react/html"
+cp "$UIPath/dist/index.html" "$ClientPath/static/react/html/"
+
+mkdir -p "$ClientPath/static/react/public"
+cp "$UIPath/public/logo.jpg" "$ClientPath/static/react/public/"
+cp "$UIPath/public/logo_1.jpg" "$ClientPath/static/react/public/"
+
+cd "$ClientPath"  exit 1
+
 go build
 if [ $? -ne 0 ]; then
     echo "Go build failed"
     exit 1
 fi
 
-# Copy the Go executable to the output path
-OutputDir="$OutputPath/cli$Port"
-mkdir -p "$OutputDir"
-cp "$ClientPath/avdol-client" "$OutputDir/" || { echo "Failed to copy avdol-client"; exit 1; }
+mkdir -p "$OutputPath/cli$Port"
+cp "$ClientPath/avdol-client" "$OutputPath/cli$Port/"
 
-# Navigate to the output directory
-cd "$OutputDir" || { echo "Output directory not found: $OutputDir"; exit 1; }
+cd "$OutputPath/cli$Port"  exit 1
 
-# Set environment variables
 export AVDOL_CLIENT_API_PORT="$Port"
 export AVDOL_CLIENT_LOCAL_DB="avdol$Port.locdb"
 export AVDOL_CLIENT_LOG_PREFIX="dmvstv_v-local"
@@ -48,8 +54,4 @@ export AVDOL_SRV_ADMIN_PORT="8082"
 export AVDOL_SRV_ADMIN_TOKEN="admin"
 export AVDOL_SRV_TCP_PORT="8084"
 
-# Run the client executable
 ./avdol-client
-if [ $? -ne 0 ]; then
-    echo "Failed to run avdol-client"
-    exit
